Account for edge cases in crash intercepting

+++ src/ts/error.ts
@@ -4,6 +4,7 @@ import { __Console__ } from "./console";
 import { Crash } from "./crash";
 import { WaveKernel } from "./kernel";
 import { ProcessHandler } from "./process/handler";
+import { Log } from "./kernel/logging";
 export function handleGlobalErrors() {
   let LOCKED = false;
   function Error(e: ErrorEvent | PromiseRejectionEvent) {
@@ -64,21 +65,18 @@ export function handleGlobalErrors() {
 export function checkIndevTpa(stack: string, e: Error): boolean {
   const parsed = stackTraceParser.parse(stack);
   const isTpa = !!parsed[0]?.file?.includes(`localhost:3128`) || !!parsed[0]?.file?.includes(`/tpa/new/`);
-  const isSvelte = stack.includes("https://svelte.dev/e/");
-
   const handler = WaveKernel?.get()?.getModule<ProcessHandler>?.("stack", true);
   const renderer = handler?.renderer;
 
-  if (renderer && renderer.lastInteract && parsed[0]?.file?.includes(`/${renderer.lastInteract.app.id}@`)) {
-    if (isTpa) {
-      renderer.notifyCrash(renderer.lastInteract.app.data, e, renderer.lastInteract);
-      handler.kill(renderer.lastInteract.pid);
-      renderer.lastInteract = undefined;
-    } else if (isSvelte) {
-      renderer.notifyCrash(renderer.lastInteract.app.data, e, renderer.lastInteract);
-      handler.kill(renderer.lastInteract.pid);
-      renderer.lastInteract = undefined;
-    }
+  console.log();
+
+  if (isTpa && renderer && renderer.lastInteract && parsed[0]?.file?.includes(`/${renderer.lastInteract.app.id}@`)) {
+    Log("checkIndevTpa", `Not crashing for ${e instanceof PromiseRejectionEvent ? e.reason : e}: source is a TPA`);
+    handler.BUSY = false;
+    handler.dispatch.dispatch("stack-not-busy");
+    renderer.notifyCrash(renderer.lastInteract.app.data, e, renderer.lastInteract);
+    handler.kill(renderer.lastInteract.pid);
+    renderer.lastInteract = undefined;
   }
 
   return isTpa;
